Rework:
Implemented a Global profile that is dynamically mapped to the individual job's profile on load.

Reworked Set Definitions to Auto-generate the file upon changing to the job, if no file already exists.

Implemented automatic generation of profiles that load the Global profile for all jobs which do not have a profile already upon running the global profile.